Fix the error “SyntaxError Unexpected Token in JSON”
I've so that from time to time, the response received from the API was a string.

As an example, whenever inside of the quoteText was a backslash, the response received from the API was a string.
e.g.:
{"quoteText":"I can\'t imagine a person becoming a success who doesn\'t give this game of life everything hes got.", "quoteAuthor":"Walter Cronkite", "senderName":"", "senderLink":"", "quoteLink":"http://forismatic.com/en/342c3061d1/"}

As a result of that, I've switched from the `fetch` API to using the Axios library and integrated additional logic in order to handle the "string" response received from the API.