Extend the RegEx to match more non-characters
Context:

I've seen that from time to time, the response received from the API was a string.

As an example, whenever inside of the quoteText was a non-word character, the response received from the API was a string. The error in the console was: "SyntaxError Unexpected Token in JSON".

e.g.:
{"quoteText":"I can\'t imagine a person becoming a success who doesn\'t give this game of life everything hes got.", "quoteAuthor":"Walter Cronkite", "senderName":"", "senderLink":"", "quoteLink":"http://forismatic.com/en/342c3061d1/"}

As a result of that, I've extended the RegEx expression to match non-word characters.